third_party/WebKit/LayoutTests/platform/mac/fast/css/vertical-align-length-copy-bug-expected.txt

Test for https://bugs.webkit.org/show_bug.cgi?id=71996

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS style.verticalAlign is "-1px"
FAIL test.offsetTop should be 10. Was 9.
PASS successfullyParsed is true

TEST COMPLETE
PASS successfullyParsed is true

TEST COMPLETE
